backend/requirements.txt
Django==5.2.4
gunicorn
psycopg2-binary==2.9.9
djangorestframework==3.15.2
drf-yasg==1.21.7
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.5.0
pillow==11.3.0
openai==1.98.0
pytesseract==0.3.13
python-dotenv==1.1.1
tellurium==2.2.11.1
SQLAlchemy
cobra
boto3
django-storages==1.14.6
